test: split domain context lifecycle coverage
Constraint: Preserve every lifecycle and privacy assertion while changing only the five assigned test paths.
Rejected: Duplicate shared fixtures across new modules | importing focused test fixtures keeps isolation deterministic and avoids helper drift.
Confidence: high
Scope-risk: narrow
Directive: Keep future domain-context scenarios in the module matching their lifecycle, isolation, or persistence responsibility.
Tested: 14 owned unittest scenarios; assertion-count parity; compileall; Ruff; git diff --check.
Not-tested: Full repository test suite; no production behavior changed.
