+++ tests/test_domain_context_lifecycle.py
@@ -15,13 +15,6 @@ from _local_package import load_local_package
 
 load_local_package()
 
-from omh.paths import resolve_paths
-from omh.wrapper_sessions import create_or_resume_wrapper_session
-from omh.workflows.domain_project_context import bind_plugin_project
-
-from test_plugin_distribution import FakeHermesContext, load_installed_plugin
-
-
 PUBLIC_CONTEXT_KEYS = {
     "schema_version",
     "workflow_hint",
@@ -34,12 +27,6 @@ REPOSITORY_ROOT = Path(__file__).resolve().parents[1]
 EN_SALES_QUESTION = (
     "Which account or customer segment should this sales work focus on?"
 )
-KO_SALES_QUESTION = (
-    "이 영업 작업은 어떤 계정 또는 고객 세그먼트에 집중해야 하나요?"
-)
-EN_FINANCE_QUESTION = "Which reporting period should this finance analysis cover?"
-
-
 def _canonical(value: object) -> str:
     return json.dumps(value, sort_keys=True, separators=(",", ":"), ensure_ascii=False)
 
@@ -186,30 +173,6 @@ def _unresolved_message(phrase: str) -> str:
     return f"something {phrase} feels unresolved"
 
 
-def _session_turn(
-    root: Path,
-    message: str,
-    *,
-    event_id: str = "event-lifecycle",
-    channel_ref: str = "channel-lifecycle",
-) -> dict[str, object]:
-    return _run_omh(
-        root,
-        "--scope",
-        "project",
-        "chat",
-        "session",
-        "start",
-        "--source",
-        "discord",
-        "--source-event-id",
-        event_id,
-        "--channel-ref",
-        channel_ref,
-        message,
-    )
-
-
 def _assert_applied_context(
     testcase: unittest.TestCase,
     interaction: dict[str, object],
@@ -241,268 +204,6 @@ def _block_external_connections():
 
 
 class DomainContextLifecycleTests(unittest.TestCase):
-    def test_public_capture_approve_replace_retire_journey(self) -> None:
-        with TemporaryDirectory() as temporary:
-            root = _repository(Path(temporary) / "redwood-project")
-            phrase = "expert-marker-100-en"
-            message = _unresolved_message(phrase)
-
-            before_approval = _chat(root, message)
-            self.assertNotIn("domain_routing_context", before_approval)
-            generic_body = before_approval["chat_response"]["body"]
-            baseline_route = _canonical(before_approval["route"])
-            baseline_candidate = _canonical(
-                before_approval["route"].get("candidate_handoff")
-            )
-
-            first_candidate_id = _capture_candidate(
-                root,
-                scope_kind="project",
-                scope_ref="redwood-project",
-                domain="redwood-operations",
-                phrase=phrase,
-                canonical="redwood_review_marker",
-                workflow_hints=("sales-development",),
-            )
-            first_approval = _approve_candidate(root, first_candidate_id)
-            self.assertEqual(first_approval["decision"], "approved")
-
-            english = _chat(root, message)
-            korean = _chat(root, f"뭔가 {phrase} 관련해서 애매해요")
-            _assert_applied_context(
-                self,
-                english,
-                workflow="sales-development",
-                locale="en",
-                question=EN_SALES_QUESTION,
-            )
-            _assert_applied_context(
-                self,
-                korean,
-                workflow="sales-development",
-                locale="ko",
-                question=KO_SALES_QUESTION,
-            )
-            self.assertEqual(_canonical(english["route"]), baseline_route)
-            self.assertEqual(
-                _canonical(english["route"].get("candidate_handoff")),
-                baseline_candidate,
-            )
-
-            setup = _run_omh(
-                root, "--scope", "project", "setup", "--with-plugin", "--json"
-            )
-            self.assertIn(setup.get("status"), {"installed", "updated", "ok", None})
-            plugin = load_installed_plugin(root / ".hermes" / "plugins" / "omh")
-            context = FakeHermesContext()
-            plugin.register(context)
-            plugin_handler = context.tools["omh_interact"]["args"][2]
-            with _block_external_connections():
-                plugin_interaction = json.loads(
-                    plugin_handler(
-                        {
-                            "message": message,
-                            "source": "discord",
-                            "record_session": False,
-                        },
-                        project_root=str(root),
-                    )
-                )
-            _assert_applied_context(
-                self,
-                plugin_interaction,
-                workflow="sales-development",
-                locale="en",
-                question=EN_SALES_QUESTION,
-            )
-
-            first_turn = _session_turn(root, message)
-            self.assertFalse(first_turn["resumed"])
-            _assert_applied_context(
-                self,
-                first_turn["interaction"],
-                workflow="sales-development",
-                locale="en",
-                question=EN_SALES_QUESTION,
-            )
-
-            replacement_id = _capture_candidate(
-                root,
-                scope_kind="project",
-                scope_ref="redwood-project",
-                domain="redwood-operations",
-                phrase=phrase,
-                canonical="redwood_review_marker",
-                workflow_hints=("finance-analysis",),
-            )
-            replacement = _approve_candidate(root, replacement_id)
-            self.assertEqual(replacement["profile"]["revision"], 2)
-            next_turn = _session_turn(root, message)
-            self.assertTrue(next_turn["resumed"])
-            self.assertEqual(
-                next_turn["session"]["session_id"], first_turn["session"]["session_id"]
-            )
-            _assert_applied_context(
-                self,
-                next_turn["interaction"],
-                workflow="finance-analysis",
-                locale="en",
-                question=EN_FINANCE_QUESTION,
-            )
-
-            retired = _run_omh(
-                root,
-                "--scope",
-                "project",
-                "memory",
-                "domain-retire",
-                "--scope-kind",
-                "project",
-                "--scope-ref",
-                "redwood-project",
-                "--domain",
-                "redwood-operations",
-                "--retired-by",
-                "lifecycle-test",
-                "--reason",
-                "superseded",
-            )
-            self.assertEqual(retired["decision"], "retired")
-            after_retirement = _chat(root, message)
-            self.assertNotIn("domain_routing_context", after_retirement)
-            self.assertEqual(after_retirement["chat_response"]["body"], generic_body)
-            self.assertEqual(_canonical(after_retirement["route"]), baseline_route)
-
-    def test_same_named_repositories_use_only_the_bound_project_store(self) -> None:
-        with TemporaryDirectory() as temporary:
-            base = Path(temporary)
-            first = _repository(base / "one" / "same-project")
-            second = _repository(base / "two" / "same-project")
-            phrase = "expert-marker-200-en"
-            _capture_and_approve(first, domain="isolation", phrase=phrase)
-
-            applied = _chat(first, _unresolved_message(phrase))
-            isolated = _chat(second, _unresolved_message(phrase))
-
-            self.assertIn("domain_routing_context", applied)
-            self.assertNotIn("domain_routing_context", isolated)
-
-    def test_user_and_organization_profiles_are_not_consumed(self) -> None:
-        with TemporaryDirectory() as temporary:
-            root = _repository(Path(temporary) / "scope-project")
-            phrase = "expert-marker-300-en"
-            _capture_and_approve(
-                root,
-                scope_kind="user",
-                scope_ref="user-lifecycle",
-                domain="user-domain",
-                phrase=phrase,
-            )
-            _capture_and_approve(
-                root,
-                scope_kind="organization",
-                scope_ref="org-lifecycle",
-                domain="organization-domain",
-                phrase=phrase,
-            )
-
-            interaction = _chat(root, _unresolved_message(phrase))
-
-            self.assertNotIn("domain_routing_context", interaction)
-
-    def test_packaged_plugin_rejects_absent_and_hostile_redirects(self) -> None:
-        with TemporaryDirectory() as temporary:
-            base = Path(temporary)
-            trusted = _repository(base / "trusted-project")
-            hostile = _repository(base / "hostile-project")
-            phrase = "expert-marker-400-en"
-            _capture_and_approve(trusted, domain="trusted", phrase=phrase)
-            _capture_and_approve(hostile, domain="hostile", phrase=phrase)
-            _run_omh(
-                trusted, "--scope", "project", "setup", "--with-plugin", "--json"
-            )
-            plugin = load_installed_plugin(trusted / ".hermes" / "plugins" / "omh")
-            context = FakeHermesContext()
-            plugin.register(context)
-            handler = context.tools["omh_interact"]["args"][2]
-            args = {
-                "message": _unresolved_message(phrase),
-                "source": "discord",
-                "record_session": False,
-                "project_root": str(hostile),
-                "omh_home": str(hostile / ".omh"),
-                "source_metadata": {"project_ref": str(hostile)},
-            }
-
-            with mock.patch.dict(
-                os.environ,
-                {
-                    "PROJECT_ROOT": str(hostile),
-                    "OMH_HOME": str(hostile / ".omh"),
-                },
-            ), _block_external_connections():
-                absent = json.loads(handler(args))
-                applied = json.loads(handler(args, project_root=str(trusted)))
-
-            self.assertNotIn("domain_routing_context", absent)
-            _assert_applied_context(
-                self,
-                applied,
-                workflow="sales-development",
-                locale="en",
-                question=EN_SALES_QUESTION,
-            )
-
-    def test_cross_repository_session_reuse_uses_each_turns_fresh_binding(self) -> None:
-        with TemporaryDirectory() as temporary:
-            base = Path(temporary)
-            first = _repository(base / "first" / "same-project")
-            second = _repository(base / "second" / "same-project")
-            phrase = "expert-marker-500-en"
-            _capture_and_approve(first, domain="first", phrase=phrase)
-            _capture_and_approve(
-                second,
-                domain="second",
-                phrase=phrase,
-                workflow_hints=("finance-analysis",),
-            )
-            shared_paths = resolve_paths(base / "shared-omh", base / "shared-hermes")
-            metadata = {"source_event_id": "same-event", "channel_ref": "same-channel"}
-
-            with _block_external_connections(), bind_plugin_project(
-                {"project_root": str(first)}
-            ) as first_binding:
-                first_turn = create_or_resume_wrapper_session(
-                    shared_paths,
-                    _unresolved_message(phrase),
-                    source="discord",
-                    source_metadata=metadata,
-                    _host_project_binding=first_binding,
-                )
-            with _block_external_connections(), bind_plugin_project(
-                {"project_root": str(second)}
-            ) as second_binding:
-                second_turn = create_or_resume_wrapper_session(
-                    shared_paths,
-                    _unresolved_message(phrase),
-                    source="discord",
-                    source_metadata=metadata,
-                    _host_project_binding=second_binding,
-                )
-
-            self.assertTrue(second_turn["resumed"])
-            self.assertEqual(
-                first_turn["session"]["session_id"], second_turn["session"]["session_id"]
-            )
-            self.assertEqual(
-                first_turn["interaction"]["domain_routing_context"]["workflow_hint"],
-                "sales-development",
-            )
-            self.assertEqual(
-                second_turn["interaction"]["domain_routing_context"]["workflow_hint"],
-                "finance-analysis",
-            )
-
     def test_unrelated_malformed_health_artifact_suppresses_context(self) -> None:
         with TemporaryDirectory() as temporary:
             root = _repository(Path(temporary) / "unhealthy-project")
